Add common Fork-Merge implementations
